database-parser/requirements.txt
PyVCF3 >= 1.0.3
tk >= 0.1.0
